% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{cvle}
\alias{cvle}
\title{Constructor for the cvle class}
\usage{
cvle(Coef.mat, Runtime, lambda, n, Met.mat, HRTrain, HRTest, pld, Mdata)
}
\arguments{
\item{Coef.mat}{A matrix of coefficients with rows equals to number of cross
validations and columns equals to number of metabolites.}

\item{Runtime}{A vector of runtime for each iteration measured in seconds.}

\item{lambda}{A vector of estimated optimum lambda for each iterations.}

\item{n}{A vector of the number of selected metabolites}

\item{Met.mat}{A matrix with 0 and 1. Number of rows equals to number of
iterations and number of columns equals to number of metabolites. 1 indicates
that the particular metabolite was selected or had nonzero coefficient and
otherwise it is zero.}

\item{HRTrain}{A matrix of survival information for the training dataset.
It has three columns representing the estimated HR, the 95\% lower confidence
interval and the 95\% upper confidence interval.}

\item{HRTest}{A matrix of survival information for the test dataset. It has
three columns representing the estimated HR, the 95\% lower confidence
interval and the 95\% upper confidence interval.}

\item{pld}{A vector of partial likelihood deviance at each cross
validations.}

\item{Mdata}{The metabolite matrix that was used for the analysis which can
either be the full the full data or a reduced supervised PCA version.}
}
\value{
object of class cvle
}
\description{
Constructor for the cvle class
}
