% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{cvmm}
\alias{cvmm}
\title{Constructor for the cvmm class}
\usage{
cvmm(HRTrain, HRTest, train, test, n.mets, Ncv, Rdata)
}
\arguments{
\item{HRTrain}{A 3-way array, The first dimension is the number of metabolites, the second dimension is the HR statistics for the low risk group in the train dataset (HR,1/HR LCI, UCI) while the third dimension is the number of cross validation performed.}

\item{HRTest}{A 3-way array, The first dimension is the number of metabolites,
the second dimension is the HR statistics for the low risk group in the test
dataset (HR,1/HR LCI, UCI) while the third dimension is the number of
cross validation performed.}

\item{train}{The selected subjects for each CV in the train dataset}

\item{test}{The selected subjects for each CV in the test dataset}

\item{n.mets}{The number of metabolite used in the analysis}

\item{Ncv}{The number of cross validation performed}

\item{Rdata}{The Metabolite data matrix that was used for the analysis either
same as Mdata or a reduced version}
}
\value{
object of class cvmm
}
\description{
Constructor for the cvmm class
}
