% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{cvmv}
\alias{cvmv}
\title{Constructor for the cvmv class}
\usage{
cvmv(HRTrain, HRTest, Ncv, Mdata, Progfact)
}
\arguments{
\item{HRTrain}{A matrix of survival information for the training dataset.
It has three columns representing the estimated HR, the 95\% lower
confidence interval and the 95\% upper confidence interval.}

\item{HRTest}{A matrix of survival information for the test dataset.
It has three columns representing the estimated HR, the 95\%
lower confidence interval and the 95\% upper confidence interval.}

\item{Ncv}{The number of cross validation used}

\item{Mdata}{The Metabolite data matrix that was used for the analysis
either same as Mdata or a reduced version.}

\item{Progfact}{The names of prognostic factors used}
}
\value{
object of class cvmv
}
\description{
Constructor for the cvmv class
}
