% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{cvpp}
\alias{cvpp}
\title{Constructor for the cvpp class}
\usage{
cvpp(Results, Ncv, Method, CVtrain, CVtest, Nmet)
}
\arguments{
\item{Results}{A dataframe containg the estimated Hazard ratio of the
test dataset and the training dataset}

\item{Ncv}{The number of cross validation performed}

\item{Method}{The dimesion reduction method used}

\item{CVtrain}{The training dataset indices matrix used for the cross
validation}

\item{CVtest}{The test dataset indices matrix used for the cross
validation}

\item{Nmet}{The number of metabolite used for the dimesion reduction
method used}
}
\value{
object of class cvpp
}
\description{
Constructor for the cvpp class
}
