% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{cvsim}
\alias{cvsim}
\title{Constructor for the cvsim class}
\usage{
cvsim(HRpca, HRpls, Nmets, Ncv, Top)
}
\arguments{
\item{HRpca}{A 3-way array in which first, second, and third
dimensions correspond to number of metabolites, Hazard ratio
infromation(Estimated HR, LowerCI and UpperCI),
and number of cross validation respectively.
This contains the estimated HR on test data and
dimension reduction method is PCA.}

\item{HRpls}{A 3-way array in which first, second, and
third dimensions correspond to number of metabolites,
Hazard ratio infromation(Estimated HR, LowerCI and UpperCI),
and number of cross validation respectively. This contains the
estimated HR on test data and dimension reduction method is
PLS.}

\item{Nmets}{The number of metabolites in the reduced matrix}

\item{Ncv}{The number of cross validation done}

\item{Top}{A sequence of top k metabolites considered.
Default is Top=seq(5,100,by=5)}
}
\value{
object of class cvsim
}
\description{
Constructor for the cvsim class
}
