% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{perm}
\alias{perm}
\title{Constructor for the perm class}
\usage{
perm(HRobs, HRperm, nperm, Validation)
}
\arguments{
\item{HRobs}{Estimated HR for low risk group on the original data.}

\item{HRperm}{Estimated HR for low risk group on the permuted data}

\item{nperm}{Number of permutations carried out.}

\item{Validation}{The validation scheme that was used.}
}
\value{
object of class perm
}
\description{
Constructor for the perm class
}
