% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{plot,cvle,missing-method}
\alias{plot,cvle,missing-method}
\title{Plot method for cvle class}
\usage{
\S4method{plot}{cvle,missing}(x, y, type = 1, ...)
}
\arguments{
\item{x}{A cvle class object}

\item{y}{missing}

\item{type}{Plot type. 1 distribution of the HR under training and test set. 2 HR vs number selected metabolites.}

\item{...}{The usual extra arguments to generic functions — see  \code{\link[graphics]{plot.default}}}
}
\value{
Cross Valdiated Results for Lasso and Elastic Net based Predictive Metabolite plots
}
\description{
Plot method for cvle class
}
