% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{plot,cvmm,ANY-method}
\alias{plot,cvmm,ANY-method}
\title{Plot method for cvmm class}
\usage{
\S4method{plot}{cvmm,ANY}(x, y, which = 1, ...)
}
\arguments{
\item{x}{A cvmm class object}

\item{y}{missing}

\item{which}{This specify which metabolite for which estimated HR
information need to be visualized. By default results of the first metabolite
is used.}

\item{...}{The usual extra arguments to generic functions — see  \code{\link[graphics]{plot.default}}}
}
\value{
Cross Valdiated Metabolic Specific CoxPh plots
}
\description{
Plot method for cvmm class
}
