% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{plot,cvmv,ANY-method}
\alias{plot,cvmv,ANY-method}
\title{Plot method for cvmv class}
\usage{
\S4method{plot}{cvmv,ANY}(x, y, ...)
}
\arguments{
\item{x}{A cvmv class object}

\item{y}{missing}

\item{...}{The usual extra arguments to generic functions — see  \code{\link[graphics]{plot.default}}}
}
\value{
Cross validation for Majority Votes Based Classification Analysis plots
}
\description{
Plot method for cvmv class
}
