% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{plot,cvpp,missing-method}
\alias{plot,cvpp,missing-method}
\title{Plot method for cvpp class}
\usage{
\S4method{plot}{cvpp,missing}(x, y, ...)
}
\arguments{
\item{x}{A cvpp class object}

\item{y}{missing}

\item{...}{The usual extra arguments to generic functions — see  \code{\link[graphics]{plot.default}}}
}
\value{
Cross Validations for PCA and PLS based plots
}
\description{
Plot method for cvpp class
}
