% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{plot,cvsim,missing-method}
\alias{plot,cvsim,missing-method}
\title{Plot method for cvsim class}
\usage{
\S4method{plot}{cvsim,missing}(x, y, type = 1, ...)
}
\arguments{
\item{x}{A cvsim class object}

\item{y}{missing}

\item{type}{Plot type. 1 distribution of the HR under test For the Top K
metabolites using PCA. 2 distribution of the HR under test For the Top K
metabolites using PLS.}

\item{...}{The usual extra arguments to generic functions — see
\code{\link[graphics]{plot.default}}}
}
\value{
Cross validation for sequentially increases metabolites plots
}
\description{
Plot method for cvsim class
}
