% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{plot,fcv,missing-method}
\alias{plot,fcv,missing-method}
\title{Plot method for fcv class}
\usage{
\S4method{plot}{fcv,missing}(x, y, type = 1, ...)
}
\arguments{
\item{x}{A fcv class object}

\item{y}{missing}

\item{type}{Plot type. 1 is the distribution of the inner cross validated
HR under test data for each outer iterations and estimated HR on the out
of bag data are superimposed. 2 Estimated HR Density for low Risk Group.}

\item{...}{The usual extra arguments to generic functions — see
\code{\link[graphics]{plot.default}}}
}
\value{
Inner and Outer Cross Validations for Lasso Elastic Net Survival predictive models and Classification plots
}
\description{
Plot method for fcv class
}
