% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{plot,ms,ANY-method}
\alias{plot,ms,ANY-method}
\title{Plot method for ms class}
\usage{
\S4method{plot}{ms,ANY}(x, y, ...)
}
\arguments{
\item{x}{A ms class object}

\item{y}{missing}

\item{...}{The usual extra arguments to generic functions — see  \code{\link[graphics]{plot.default}}}
}
\value{
Metabolite by Metabolite CoxPh plots
}
\description{
Plot method for ms class
}
