% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{plot,perm,ANY-method}
\alias{plot,perm,ANY-method}
\title{Plot method for perm class}
\usage{
\S4method{plot}{perm,ANY}(x, y, ...)
}
\arguments{
\item{x}{A perm class object}

\item{y}{missing}

\item{...}{The usual extra arguments to generic functions — see
\code{\link[graphics]{plot.default}}}
}
\value{
Null Distribution of the Estimated HR plots
}
\description{
Plot method for perm class
}
