% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{show,cvmm-method}
\alias{show,cvmm-method}
\title{Show method for cvmm class}
\usage{
\S4method{show}{cvmm}(object)
}
\arguments{
\item{object}{An object of class cvmm}
}
\value{
Cross Valdiated Metabolic Specific CoxPh information
}
\description{
Show method for cvmm class
}
