% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{show,cvmv-method}
\alias{show,cvmv-method}
\title{Show method for cvmv class}
\usage{
\S4method{show}{cvmv}(object)
}
\arguments{
\item{object}{An object of class cvmv}
}
\value{
Cross validation for Majority Votes Based Classification Analysis information
}
\description{
Show method for cvmv class
}
