% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{show,cvpp-method}
\alias{show,cvpp-method}
\title{Show method for cvpp class}
\usage{
\S4method{show}{cvpp}(object)
}
\arguments{
\item{object}{An object of class cvpp}
}
\value{
CCross Validations for PCA and PLS based information
}
\description{
Show method for cvpp class
}
