% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{show,ms-method}
\alias{show,ms-method}
\title{Show method for ms class}
\usage{
\S4method{show}{ms}(object)
}
\arguments{
\item{object}{An object of class ms}
}
\value{
Metabolite by Metabolite CoxPh Model and the number of
metabolites used.
}
\description{
Show method for ms class
}
