% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{show,perm-method}
\alias{show,perm-method}
\title{Show method for perm class}
\usage{
\S4method{show}{perm}(object)
}
\arguments{
\item{object}{An object of class perm}
}
\value{
Null Distribution of the Estimated HR information
}
\description{
Show method for perm class
}
