% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{summary,cvle-method}
\alias{summary,cvle-method}
\title{Summary method for cvle class}
\usage{
\S4method{summary}{cvle}(object)
}
\arguments{
\item{object}{An object of class cvle}
}
\value{
Cross Valdiated Results for Lasso and Elastic Net based Predictive Metabolite signature summary information
}
\description{
Summary method for cvle class
}
