% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{summary,cvmm-method}
\alias{summary,cvmm-method}
\title{Summary method for cvmm class}
\usage{
\S4method{summary}{cvmm}(object, which = 1)
}
\arguments{
\item{object}{An object of class cvmm}

\item{which}{This specify which metabolite for which estimated HR}
}
\value{
Cross Valdiated Metabolic Specific CoxPh summary
}
\description{
Summary method for cvmm class
}
