% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{summary,cvmv-method}
\alias{summary,cvmv-method}
\title{Summary method for cvmv class}
\usage{
\S4method{summary}{cvmv}(object)
}
\arguments{
\item{object}{An object of class cvmv}
}
\value{
Cross validation for Majority Votes Based Classification Analysis summary
}
\description{
Summary method for cvmv class
}
