% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{summary,cvpp-method}
\alias{summary,cvpp-method}
\title{Summary method for cvpp class}
\usage{
\S4method{summary}{cvpp}(object)
}
\arguments{
\item{object}{An object of class cvpp}
}
\value{
Cross Validations for PCA and PLS based summary
}
\description{
Summary method for cvpp class
}
