% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{summary,cvsim-method}
\alias{summary,cvsim-method}
\title{Summary method for cvsim class}
\usage{
\S4method{summary}{cvsim}(object)
}
\arguments{
\item{object}{An object of class cvsim}
}
\value{
Cross validation for sequentially increases metabolites summary
}
\description{
Summary method for cvsim class
}
