% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{summary,fcv-method}
\alias{summary,fcv-method}
\title{Summary method for fcv class}
\usage{
\S4method{summary}{fcv}(object)
}
\arguments{
\item{object}{An object of class fcv}
}
\value{
Inner and Outer Cross Validations for Lasso Elastic Net Survival predictive models and Classification summary
}
\description{
Summary method for fcv class
}
