% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{summary,ms-method}
\alias{summary,ms-method}
\title{Summary method for ms class}
\usage{
\S4method{summary}{ms}(object)
}
\arguments{
\item{object}{An object of class ms}
}
\value{
Metabolite by Metabolite CoxPh summary information
}
\description{
Summary method for ms class
}
