% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicSurvClass.R
\name{summary,perm-method}
\alias{summary,perm-method}
\title{Summary method for perm class}
\usage{
\S4method{summary}{perm}(object)
}
\arguments{
\item{object}{An object of class perm}
}
\value{
Null Distribution of the Estimated HR summary
}
\description{
Summary method for perm class
}
